Rename

+++ PART II/AsianCallOption.cpp
@@ -8,7 +8,7 @@ OptionType AsianCallOption::GetOptionType() const { return OptionType::Call; }
 
 double AsianCallOption::payoffPath(vector<double>& past_prices) const {
 	double price = 0;
-	for (int i = 0; i < past_prices.size(); i++) {
+	for (size_t i = 0; i < past_prices.size(); i++) {
 		price += past_prices[i];
 	}
 	if (price / past_prices.size() >= GetStrike()) {
