+++ PART II/AsianPutOption.cpp
@@ -8,7 +8,7 @@ OptionType AsianPutOption::GetOptionType() const { return OptionType::Put; }
 
 double AsianPutOption::payoffPath(vector<double>& past_prices) const {
 	double price = 0;
-	for (int i = 0; i < past_prices.size(); i++) {
+	for (size_t i = 0; i < past_prices.size(); i++) {
 		price += past_prices[i];
 	}
 	if (price / past_prices.size() <= GetStrike()) {
